Rename folder, conflicting on windows
