+++ tests/Functional/Controller/BaseAclControllerTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Packeton\Tests\Functional\Controller;
 
-use Packeton\Tests\Phpunit\PacketonTestTrait;
+use Packeton\Tests\Functional\PacketonTestTrait;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
