+++ tests/phpunit
@@ -1,6 +1,6 @@
 #!/usr/bin/env php
 <?php
 
-require_once __DIR__ . '/Phpunit/ErrorHandler.php';
+require_once __DIR__ . '/PHPUnitOverride/ErrorHandler.php';
 
 include dirname(__DIR__) . '/vendor/bin/phpunit';
